requirements.txt
# Core AI and ML libraries with CUDA support
torch>=2.1.0+cu121
torchvision>=0.16.0+cu121
torchaudio>=2.1.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# Advanced AI models for lipsync (optimized for 16GB VRAM)
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
xformers>=0.0.22
bitsandbytes>=0.41.0
flash-attn>=2.3.0
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
scikit-image>=0.21.0
face-recognition>=1.3.0
mediapipe>=0.10.0
insightface>=0.7.3
onnxruntime-gpu>=1.16.0

# Audio Processing (high-quality for real-time)
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
elevenlabs>=0.2.0
espnet>=202301
resampy>=0.4.2
pyworld>=0.3.0
praat-parselmouth>=0.4.3

# Web Framework and API (production-ready)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
websockets>=12.0.0
sse-starlette>=1.6.0
redis>=5.0.0
celery>=5.3.0

# Utilities (performance optimized)
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0
psutil>=5.9.0
gpustat>=1.1.0
nvidia-ml-py>=12.0.0

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0
av>=10.0.0

# Additional ML utilities
einops>=0.7.0
omegaconf>=2.3.0
safetensors>=0.4.0
huggingface-hub>=0.17.0
timm>=0.9.0
controlnet-aux>=0.0.6
